requirements.txt
streamlit
openai
anthropic
python-docx
PyMuPDF